requirements.txt
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0

fastapi==0.68.0
uvicorn==0.15.0
line-bot-sdk==1.20.0
